+++ common/src/main/java/com/github/cao/awa/annuus/information/compressor/inaction/InactionCompressor.java
@@ -10,6 +10,11 @@ public class InactionCompressor implements InformationCompressor {
     }
 
     @Override
+    public String getAlias() {
+        return null;
+    }
+
+    @Override
     public int getId() {
         return -1;
     }